Port autograph integration test to python
